+++ resources/views/event/admin/overview.blade.php
@@ -66,6 +66,10 @@
                                 class="fa fa-envelope"></i> Send E-mails (all persons)</a>&nbsp;
                         <a href="{{ route('admin.bookings.export',$event) }}" class="btn btn-success"><i
                                 class="fa fa-edit"></i> Export data</a>&nbsp;
+                        @if($event->event_type_id == \App\Enums\EventType::MULTIFLIGHTS)
+                            <a href="{{ route('admin.bookings.export',[$event, 'vacc']) }}" class="btn btn-success"><i
+                                    class="fa fa-edit"></i> Export data + VACC data</a>&nbsp;
+                        @endif
                         @if($event->startEvent > now())
                             <form action="{{ route('admin.events.destroy', $event) }}" method="post">
                                 @csrf