+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.5.1](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v1.5.1...v1.5.0.c) - 2020-01-15
+
 ### Added
 - Export for Multi-city event type (with e-mails and with CTOT's)
 - Added ``auth.isLoggedIn`` middleware to ``BookingController@edit()`` that you are logged in